+++ core/src/lib.rs
@@ -9,6 +9,7 @@ pub type WasmBufferPtr = WasmPtr<u8, Array>;
 pub mod abi;
 pub mod buffers;
 pub mod threader;
+pub mod wasm;
 
 #[inline(always)]
 /// Invoke an IOmod call at coordinates `method_path` with input `method_input`
