+++ providers/aws-lambda/host/Cargo.toml
@@ -21,9 +21,6 @@ toml = "0.5"
 zip = "0.5"
 
 wasmer = "2.1.1"
-wasmer-wasi = "2.1.1"
-#wasmer-engine-native = "1.0"
-wasmer-engine-universal = "2.1.1"
 
 assemblylift_core = { version = "0.3", package = "assemblylift-core", path = "../../../core" }
 assemblylift_core_iomod = { version = "0.3", package = "assemblylift-core-iomod", path = "../../../core/iomod" }
